+++ lib/providers/providers.dart
@@ -1,4 +1,4 @@
 import 'package:comics_center/download/downloader.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-final downloaderProvider = NotifierProvider<Downloader, Map<String, int>>(Downloader.new);
+final downloaderProvider = NotifierProvider<Downloader, Map<String, double>>(Downloader.new);
